backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
sarvamai
httpx==0.28.1
python-multipart==0.0.20
Pillow==11.1.0
